all component files typescript
