+++ package.json
@@ -59,6 +59,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/preset-react": "^7.18.6",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.34",
@@ -70,6 +71,7 @@
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "babel-plugin-styled-components": "^2.0.7",
+    "babel-plugin-typescript-to-proptypes": "^2.1.0",
     "eslint": "^8.53.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
