+++ src/App.test.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { Provider } from "react-redux";
-import { store } from "./app/store";
 import App from "./app/App";
 
 test("renders learn react link", () => {
   const { getByText } = render(
-    <Provider store={store}>
-      <App />
-    </Provider>
+    // <Provider store={store}>
+    <App />
+    // </Provider>
   );
 
   expect(getByText(/learn/i)).toBeInTheDocument();
