+++ src/components/form-components/InputFormItem1.tsx
@@ -1,7 +1,5 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-import { useFieldArray, useFormContext } from "react-hook-form";
-import { v4 as uuidv4 } from "uuid";
+import { useFieldArray, useFormContext} from "react-hook-form";
+// import { v4 as uuidv4 } from "uuid";
 import React, { useEffect, useRef } from "react";
 import PropTypes from "prop-types";
 import {
@@ -16,28 +14,31 @@ import {
 import NewItemButton from "../buttons/NewItemButton";
 import useWindowWidth from "../../hooks/useWindowWidth";
 import {Col, Col1} from "../../styles/EditFormItemListStyles";
-import {Invoice} from "../../types/types";
+import { Invoice} from "../../types/types";
 
 type InputFormItemProps = {
-  invoice: Invoice;
+  invoice?: Invoice;
   isDraft: boolean;
   isEditOpen: boolean;
 }
+
 export default function InputFormItem1({ isDraft, invoice, isEditOpen }: InputFormItemProps) {
 
-  const { formState, register, watch, clearErrors, setError } = useFormContext();
+  const { formState, register, watch, clearErrors, setError, resetField } = useFormContext();
 
   const { fields, remove, append } = useFieldArray({
     name: "items", rules: { required: true, minLength: 1 }
   });
 
   const { errors, isSubmitting} = formState;
+
+
   const watchItems = watch("items", []);
   const watcher = watch();
   const width = useWindowWidth();
   const isInitialRender = useRef(true);
   // eslint-disable-next-line no-console
-  console.log(errors);
+  // console.log(errors);
 
 
 
@@ -59,9 +60,20 @@ export default function InputFormItem1({ isDraft, invoice, isEditOpen }: InputFo
 
   useEffect(() => {
     if (invoice && isEditOpen) {
+      // const {items} = getValues();
+      // console.log(items);
       invoice.items.forEach(i => {
-        append({id: i.id || uuidv4(), name: i.name, quantity: i.quantity, price: i.price, total: i.total});
+        // if (!items.includes(i)) {
+        append({id: i.id, name: i.name, quantity: i.quantity, price: i.price, total: i.total});
+        // }
       });
+    }
+
+    if (!isEditOpen) {
+      setTimeout(() => {
+        resetField("items");
+      }, 200);
+
     }
     // console.log("items added to form");
   }
@@ -85,21 +97,21 @@ export default function InputFormItem1({ isDraft, invoice, isEditOpen }: InputFo
           placeholder="Item name"
           defaultValue={invoice? invoice?.items?.[index]?.name : "" }
           type="text"
-          style={{ border: errors?.items?.[index]?.name ? "1px solid #EC5757" : "" }}
+          style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.name ? "1px solid #EC5757" : "" }}
         />
       </Box>
       <SmallBoxContainer>
         <Box>
           <Col style={{ marginBottom: "0.625rem" }}>Qty.</Col>
           <Quantity {...register(`items[${index}].quantity`, { required: !isDraft, max: 100 } )} placeholder="0" type="text"
-            style={{ border: errors?.items?.[index]?.quantity ? "1px solid #EC5757" : "" }}
+            style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.quantity ? "1px solid #EC5757" : "" }}
             defaultValue={invoice? invoice?.items?.[index]?.quantity : "" }/>
         </Box>
         <Box>
           <Col style={{ marginBottom: "0.625rem" }}>Price</Col>
           <Price {...register(`items[${index}].price`, { required: !isDraft, max: 100000 } )} placeholder="0.00" type="text"
             defaultValue={invoice? invoice?.items?.[index]?.price : "" }
-            style={{ border: errors?.items?.[index]?.price ? "1px solid #EC5757" : "" }} />
+            style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.price ? "1px solid #EC5757" : "" }} />
 
         </Box>
         <TotalBox style={{ width: "fit-content" }}>
@@ -128,16 +140,16 @@ export default function InputFormItem1({ isDraft, invoice, isEditOpen }: InputFo
           placeholder="Item name"
           defaultValue={invoice? invoice?.items?.[index]?.name : "" }
           type="text"
-          style={{ border: errors?.items?.[index]?.name ? "1px solid #EC5757" : "" }}
+          style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.name ? "1px solid #EC5757" : "" }}
         />
         <Quantity {...register(`items[${index}].quantity`, { required: !isDraft, max: 100 } )} placeholder="0" type="text"
-          style={{ border: errors?.items?.[index]?.quantity ? "1px solid #EC5757" : "" }}
+          style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.quantity ? "1px solid #EC5757" : "" }}
           defaultValue={invoice? invoice?.items?.[index]?.quantity : "" }/>
 
 
         <Price {...register(`items[${index}].price`, { required: !isDraft, max: 100000 } )} placeholder="0.00" type="text"
           defaultValue={invoice? invoice?.items?.[index]?.price : "" }
-          style={{ border: errors?.items?.[index]?.price ? "1px solid #EC5757" : "" }} />
+          style={{ border: Array.isArray(errors.items) && errors?.items?.[index]?.price ? "1px solid #EC5757" : "" }} />
 
         <Total>
           { (Number(watchItems?.[index]?.quantity) * Number(watchItems?.[index]?.price)).toFixed(2)}
@@ -162,7 +174,7 @@ export default function InputFormItem1({ isDraft, invoice, isEditOpen }: InputFo
           </li>
         ))}
       </ul>
-      <NewItemButton append={append} items={invoice.items}/>
+      <NewItemButton append={append} items={invoice ? invoice.items : []}/>
     </>
   );
 }
@@ -175,6 +187,6 @@ InputFormItem1.defaultProps = {
 InputFormItem1.propTypes = {
   isDraft: PropTypes.bool.isRequired,
   isEditOpen: PropTypes.bool,
-  invoice: PropTypes.instanceOf(Invoice)
+  // invoice: PropTypes.instanceOf(Invoice)
 };
