+++ src/features/invoices/invoicesSlice.ts
@@ -15,11 +15,13 @@ export const invoicesSlice = createSlice({
       state.data.push(action.payload);
     },
     removeInvoice: (state, action) => {
+
       state.data = state.data.filter(
         (invoice) => invoice.id !== action.payload
       );
     },
     updateInvoice: (state, action: PayloadAction<Invoice>) => {
+      console.log("entered updateInvoice");
       const oldInvoice: Invoice | undefined = state.data.find(
         (invoice: Invoice) => invoice.id === action.payload.id
       );
@@ -27,6 +29,7 @@ export const invoicesSlice = createSlice({
       if (oldInvoice) {
         const index = state.data.indexOf(oldInvoice);
         state.data.splice(index, 1, action.payload);
+        console.log("invoice updated");
       }},
     markAsPaid: (state, action) => {
       const invoice1 = state.data.find(
@@ -43,14 +46,17 @@ export const invoicesSlice = createSlice({
         invoice1.status = "pending";
       }},
     addItem: (state, action) => {
+
       const { id, newItem } = action.payload;
-      const invoice = state.data.find((invoice) => invoice.id === id);
+      const invoice = state.data.find((element) => element.id === id);
       if (invoice) {
         invoice.items.push(newItem);
       }},
     addIdToExistingInvoices: (state) => {
+      // eslint-disable-next-line no-restricted-syntax
       for (const i of state.data) {
         let item: WritableDraft<Item>;
+        // eslint-disable-next-line no-restricted-syntax
         for (item of i.items) {
           if (!item.id) {
             item.id = uuidv4();
@@ -58,7 +64,7 @@ export const invoicesSlice = createSlice({
         }
       }
     },
-    clearInvoices: (state, action) => {
+    clearInvoices: (state) => {
       state.data = [];
     },
   },
