--- styled.d.ts
@@ -1,16 +1,0 @@
-// // import original module declarations
-// import 'styled-components';
-//
-//
-// // and extend them!
-// declare module 'styled-components' {
-//     export interface DefaultTheme {
-//         borderRadius: string;
-//
-//
-//         colors: {
-//             main: string;
-//             secondary: string;
-//         };
-//     }
-// }
